Adding standard methods for identifying things in substrate ads
